serenity/requirements.txt
Django==3.2.13
django-haystack==3.2.1
elasticsearch==7.13.1
gTTS==2.5.3
openai==0.28.0
python-dotenv==1.0.1
pillow==10.3.0
pygame==2.6.0